finished CRUD for artists, genres, songs

+++ tunaapi/models/song.py
@@ -4,5 +4,5 @@ from .artist import Artist
 class Song(models.Model):
     title = models.CharField(max_length=200)
     album = models.CharField(max_length=55)
-    length = models.DecimalField(max_digits=4,decimal_places=2)
+    length = models.IntegerField()
     artist_id = models.ForeignKey(Artist, on_delete=models.CASCADE)
